cores: arduino: Split zephyrCommon.cpp
Split zephyrCommon.cpp into smaller source files.

wiring_digital.cpp:
  - pinMode
  - digitalWrite
  - digitalRead

wiring_analog.cpp:
  - analogRead
  - analogWrite
  - analogReference
  - analogReadResolution
  - analogWriteResolution

wiring_pulse.cpp:
  - pulseIn

WInterrupts.cpp:
  - attachInterrupt
  - detachInterrupt
  - interrupts
  - noInterrupts
  - digitalPinToInterrupt

WMath.cpp:
  - randomSeed
  - random

Tone.cpp:
  - tone
  - noTone
